[#31] Implement the basic locator
The basic locator is implemented in class `Locator`. This locator does
**NOT** consider thickness preserving, medial transformations, consonant
replacements, region competings, or etc. Theses features should be
implemented later.

This locator do only mapping each character to its *base* region by
really simple linear transformation equations. If the letter is given by
(initial, medial, final), the *base* region of each character of it is
given by the region of each character of (Mieum, medial, Mieum),
respectively.

To deal with regions, the class `Region` is implemented. It represents a
region by its minimum and maximum values of x- and y-axis.